--- a/berkeley-function-call-leaderboard/bfcl/constants/eval_config.py
+++ b/berkeley-function-call-leaderboard/bfcl/constants/eval_config.py
@@ -69,6 +69,7 @@ UNDERSCORE_TO_DOT = [
     "NousResearch/Hermes-2-Theta-Llama-3-70B",
     "command-r-plus-FC",
     "command-r7b-12-2024-FC",
+    "command-a-03-2025-FC",
     "THUDM/glm-4-9b-chat",
     "ibm-granite/granite-20b-functioncalling",
     "yi-large-fc",