--- a/berkeley-function-call-leaderboard/bfcl/eval_checker/model_metadata.py
+++ b/berkeley-function-call-leaderboard/bfcl/eval_checker/model_metadata.py
@@ -590,6 +590,12 @@ MODEL_METADATA_MAPPING = {
         "Cohere",
         "cc-by-nc-4.0",
     ],
+    "command-a-03-2025-FC": [
+        "Command A (FC)",
+        "https://cohere.com/blog/command-a",
+        "Cohere",
+        "CC-BY-NC 4.0 License (w/ Acceptable Use Addendum)",
+    ],
     "snowflake/arctic": [
         "Snowflake/snowflake-arctic-instruct (Prompt)",
         "https://huggingface.co/Snowflake/snowflake-arctic-instruct",
@@ -1003,6 +1009,7 @@ INPUT_PRICE_PER_MILLION_TOKEN = {
     "databricks-dbrx-instruct": 2.25,
     "command-r-plus-FC": 3,
     "command-r7b-12-2024-FC": 0.0375,
+    "command-a-03-2025-FC": 2.5,
     "yi-large-fc": 3,
     "palmyra-x-004": 5,
     "grok-beta": 5,
@@ -1071,6 +1078,7 @@ OUTPUT_PRICE_PER_MILLION_TOKEN = {
     "databricks-dbrx-instruct": 6.75,
     "command-r-plus-FC": 15,
     "command-r7b-12-2024-FC": 0.15,
+    "command-a-03-2025-FC": 10,
     "yi-large-fc": 3,
     "palmyra-x-004": 12,
     "grok-beta": 15,