Hello from Cohere 👋 

This PR submits our latest model - [Command A](https://cohere.com/blog/command-a). When I ran this against [f81063](https://github.com/ShishirPatil/gorilla/commit/f8106312357d2e687a905610fe972b64482ab8e2) I got the following score, and hopefully you'll be able to reproduce them:

```
| Overall Acc | Model          | Model Link                        | Cost ($ Per 1k Function Calls) | Latency Mean (s) | Latency Standard Deviation (s) | Latency 95th Percentile (s) | Non-Live AST Acc | Non-Live Simple AST | Non-Live Multiple AST | Non-Live Parallel AST | Non-Live Parallel Multiple AST | Non-Live Exec Acc | Non-Live Simple Exec | Non-Live Multiple Exec | Non-Live Parallel Exec | Non-Live Parallel Multiple Exec | Live Acc | Live Simple AST | Live Multiple AST | Live Parallel AST | Live Parallel Multiple AST | Multi Turn Acc | Multi Turn Base | Multi Turn Miss Func | Multi Turn Miss Param | Multi Turn Long Context | Relevance Detection | Irrelevance Detection | Organization |
|-------------|----------------|-----------------------------------|--------------------------------|------------------|--------------------------------|-----------------------------|------------------|---------------------|-----------------------|-----------------------|--------------------------------|-------------------|----------------------|------------------------|------------------------|---------------------------------|----------|-----------------|-------------------|-------------------|----------------------------|----------------|-----------------|----------------------|-----------------------|-------------------------|---------------------|-----------------------|--------------|
| 63.14%      | Command A (FC) | https://cohere.com/blog/command-a | 6.74                           | 6.22             | 11.38                          | 17.1                        | 86.73%           | 73.92%              | 93.50%                | 93.50%                | 86.00%                         | 82.12%            | 97.00%               | 88.00%                 | 76.00%                 | 67.50%                          | 80.14%   | 83.72%          | 76.92%            | 81.25%            | 62.50%                     | 24.38%         | 32.00%          | 23.00%               | 22.00%                | 20.50%                  | 72.22%              | 86.16%                | Cohere       |

```

Thanks in advance for taking the time to review, and I look forward to seeing us on the leaderboard 🎉